refactor: implement access for adding and deleting mates and states in projects and tasks in the use case.

+++ src/main/kotlin/domain/usecase/project/DeleteProjectUseCase.kt
@@ -1,5 +1,6 @@
 package org.example.domain.usecase.project
 
+import org.example.domain.AccessDeniedException
 import org.example.domain.entity.log.DeletedLog
 import org.example.domain.entity.log.Log
 import org.example.domain.repository.LogsRepository
@@ -13,15 +14,18 @@ class DeleteProjectUseCase(
     private val usersRepository: UsersRepository,
 ) {
     operator fun invoke(projectId: UUID) =
-        projectsRepository.getProjectById(projectId).let { project ->
-            projectsRepository.deleteProjectById(projectId)
-            logsRepository.addLog(
-                DeletedLog(
-                    username = usersRepository.getCurrentUser().username,
-                    affectedId = projectId,
-                    affectedName = project.name,
-                    affectedType = Log.AffectedType.PROJECT,
+        usersRepository.getCurrentUser().let { currentUser ->
+            projectsRepository.getProjectById(projectId).let { project ->
+                if (project.createdBy != currentUser.id) throw AccessDeniedException("project")
+                projectsRepository.deleteProjectById(projectId)
+                logsRepository.addLog(
+                    DeletedLog(
+                        username = currentUser.username,
+                        affectedId = projectId,
+                        affectedName = project.name,
+                        affectedType = Log.AffectedType.PROJECT,
+                    )
                 )
-            )
+            }
         }
 }
